+++ install/setup.sh
@@ -6,20 +6,25 @@
 #
 
 INSTALL_DIRECTORY=$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." && pwd -P )
+CURRENT_VERSION=$(${INSTALL_DIRECTORY}/Mycodo/env/bin/python3 ${INSTALL_DIRECTORY}/Mycodo/mycodo/utils/github_release_info.py -c 2>&1)
 INSTALL_CMD="/bin/bash ${INSTALL_DIRECTORY}/mycodo/scripts/upgrade_commands.sh"
+INSTALL_DEP="/bin/bash ${INSTALL_DIRECTORY}/mycodo/scripts/dependencies.sh"
 LOG_LOCATION=${INSTALL_DIRECTORY}/install/setup.log
+INSTALL_TYPE="Full"
 
 if [ "$EUID" -ne 0 ]; then
     printf "Please run as root: \"sudo /bin/bash ${INSTALL_DIRECTORY}/install/setup.sh\"\n";
     exit
 fi
 
-exec > >(tee -i -a ${LOG_LOCATION})
-exec 2>&1
+command -v whiptail >/dev/null || {
+    printf "\nwhiptail not installed. Install it with 'sudo apt-get install whiptail' then try the install again.\n"
+    exit;
+}
 
 abort()
 {
-    echo >&2 '
+    echo '
 ************************************
 ** ERROR: Mycodo Install Aborted! **
 ************************************
@@ -27,12 +32,14 @@ abort()
 An error occurred that may have prevented Mycodo from
 being installed properly!
 
+Open to the end of the setup log to view the full error:
+${INSTALL_DIRECTORY}/install/setup.log
+
 Please contact the developer by submitting a bug report
 at https://github.com/kizniche/Mycodo/issues with the
 pertinent excerpts from the setup log located at:
 ${INSTALL_DIRECTORY}/install/setup.log
-'
-    echo "An error occurred. Exiting..." >&2
+' 2>&1 | tee -a ${LOG_LOCATION}
     exit 1
 }
 
@@ -41,203 +48,177 @@ trap 'abort' 0
 set -e
 
 NOW=$(date +"%m-%d-%Y %H:%M:%S")
-printf "### Mycodo installation began at $NOW\n"
+printf "### Mycodo installation began at $NOW\n" >>${LOG_LOCATION}
 
-INSTALL_TYPE="FULL"
+clear
+LICENSE=$(whiptail --title "Mycodo Installer: License Agreement" \
+                   --backtitle "Mycodo ${CURRENT_VERSION}" \
+                   --yesno "Mycodo is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.\nMycodo is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with Mycodo. If not, see gnu.org/licenses.\n\nDo you agree to the license terms?" \
+                   18 68 \
+                   3>&1 1>&2 2>&3)
 
-printf "\nSelect the type of install you would like to perform."
-printf "\nA 'full' install is recommended unless you know what you're doing.\n\n"
+exitstatus=$?
+if [ $exitstatus != 0 ]; then
+    echo "Install canceled by user" >>${LOG_LOCATION}
+    exit
+fi
 
-PS3="Please choose an option (1-3): "
-select option in minimum custom full
-do
-    case $option in
-        minimum)
-            INSTALL_TYPE="MINIMAL"
-            break;;
-        custom)
-            INSTALL_TYPE="CUSTOM"
-            break;;
-        full)
-            INSTALL_TYPE="FULL"
-            break;;
-     esac
-done
+clear
+INSTALL_TYPE=$(whiptail --title "Mycodo Installer: Install Type" \
+                        --backtitle "Mycodo ${CURRENT_VERSION}" \
+                        --notags \
+                        --menu "Select the Install Type:\n\nFull: Install all dependencies\nMinimal: Install a minimal set of dependencies\nCustom: Select which dependencies to install\n\nIf unsure, choose 'Full Install'" \
+                        18 68 3 \
+                        "full" "Full Install (recommended)" \
+                        "minimal" "Minimal Install" \
+                        "custom" "Custom Install" \
+                        3>&1 1>&2 2>&3)
+exitstatus=$?
+if [ $exitstatus != 0 ]; then
+    echo "Install canceled by user" >>${LOG_LOCATION}
+    exit
+fi
+printf "\nInstall Type: $INSTALL_TYPE\n"
 
-if [ "$INSTALL_TYPE" == "CUSTOM" ]; then
-    choice () {
-        local choice=$1
-        if [[ ${opts[choice]} ]] # toggle
-        then
-            opts[choice]=
-        else
-            opts[choice]=+
-        fi
-    }
+if [ "$INSTALL_TYPE" == "custom" ]; then
+    clear
+    INSTALL_DEP=$(whiptail --title "Mycodo Install: Custom" \
+                           --backtitle "Mycodo ${CURRENT_VERSION}" \
+                           --notags \
+                           --checklist "Dependencies to Install" \
+                           18 68 11 \
+                           1 "Adafruit_ADS1x15" off \
+                           2 "Adafruit_BME280" off \
+                           3 "Adafruit_GPIO" off \
+                           4 "Adafruit_MCP3008" off \
+                           5 "Adafruit_TMP" off \
+                           6 "MCP342x" off \
+                           7 "pigpio" off \
+                           8 "sht_sensor" off \
+                           9 "tsl2561" off \
+                           10 "tsl2591" off \
+                           11 "w1thermsensor" off \
+                           3>&1 1>&2 2>&3)
+    exitstatus=$?
+    if [ $exitstatus != 0 ]; then
+        echo "Install canceled by user" >>${LOG_LOCATION}
+        exit
+    fi
+    printf "\nDependencies: $INSTALL_DEP\n" >>${LOG_LOCATION}
+fi
 
-    PS3='Select which dependencies to install, then select "Done": '
-    while :
-    do
-        clear
-        options=(
-            "Adafruit_ADS1x15 ${opts[1]}"
-            "Adafruit_BME280 ${opts[2]}"
-            "Adafruit_GPIO ${opts[3]}"
-            "Adafruit_MCP3008 ${opts[4]}"
-            "Adafruit_TMP ${opts[5]}"
-            "MCP342x ${opts[6]}"
-            "pigpio ${opts[7]}"
-            "sht_sensor ${opts[8]}"
-            "tsl2561 ${opts[9]}"
-            "tsl2591 ${opts[10]}"
-            "w1thermsensor ${opts[11]}"
-            "Done"
-        )
-        select opt in "${options[@]}"
+{
+    echo -e "XXX\n4\nChecking swap size... \nXXX"
+    ${INSTALL_CMD} update-swap-size >>${LOG_LOCATION}
+
+    echo -e "XXX\n5\nUpdating apt sources... \nXXX"
+    ${INSTALL_CMD} update-apt >>${LOG_LOCATION}
+
+    echo -e "XXX\n4\nRemoving apt- version of pip... \nXXX"
+    ${INSTALL_CMD} uninstall-apt-pip >>${LOG_LOCATION}
+
+    echo -e "XXX\n5\nInstall dependencies (apt)... \nXXX"
+    ${INSTALL_CMD} update-packages >>${LOG_LOCATION}
+
+    echo -e "XXX\n4\nSetting up virtualenv... \nXXX"
+    ${INSTALL_CMD} setup-virtualenv >>${LOG_LOCATION}
+
+    echo -e "XXX\n5\nUpdating pip... \nXXX"
+    ${INSTALL_CMD} update-pip3 >>${LOG_LOCATION}
+
+    echo -e "XXX\n4\nInstalling WiringPi... \nXXX"
+    ${INSTALL_CMD} update-wiringpi >>${LOG_LOCATION}
+
+    echo -e "XXX\n5\nInstalling base python packages... \nXXX"
+    ${INSTALL_CMD} update-pip3-packages >>${LOG_LOCATION}
+
+    if [ "$INSTALL_TYPE" == "minimal" ]; then
+        printf '\n### Minimal install selected. No more dependencies to install.'
+    elif [ "$INSTALL_TYPE" == "custom" ]; then
+        printf '\n### Installing custom-selected dependencies'
+        echo -e "XXX\n4\nInstalling custom python packages... \nXXX"
+        for option in $INSTALL_DEP
         do
-            case $opt in
-                "Adafruit_ADS1x15 ${opts[1]}")
-                    choice 1
-                    break
-                    ;;
-                "Adafruit_BME280 ${opts[2]}")
-                    choice 2
-                    break
-                    ;;
-                "Adafruit_GPIO ${opts[3]}")
-                    choice 3
-                    break
-                    ;;
-                "Adafruit_MCP3008 ${opts[4]}")
-                    choice 4
-                    break
-                    ;;
-                "Adafruit_TMP ${opts[5]}")
-                    choice 5
-                    break
-                    ;;
-                "MCP342x ${opts[6]}")
-                    choice 6
-                    break
-                    ;;
-                "pigpio ${opts[7]}")
-                    choice 7
-                    break
-                    ;;
-                "sht_sensor ${opts[8]}")
-                    choice 8
-                    break
-                    ;;
-                "tsl2561 ${opts[9]}")
-                    choice 9
-                    break
-                    ;;
-                "tsl2591 ${opts[10]}")
-                    choice 10
-                    break
-                    ;;
-                "w1thermsensor ${opts[11]}")
-                    choice 11
-                    break
-                    ;;
-                "Done")
-                    break 2
-                    ;;
-                *) printf '\nInvalid option';;
-            esac
-        done
-    done
-fi
-
-${INSTALL_CMD} update-swap-size
-
-${INSTALL_CMD} update-apt
-
-${INSTALL_CMD} uninstall-apt-pip
-
-${INSTALL_CMD} update-packages
-
-${INSTALL_CMD} setup-virtualenv
-
-${INSTALL_CMD} update-pip3
-
-${INSTALL_CMD} update-wiringpi
-
-${INSTALL_CMD} update-pip3-packages
-
-if [ "$INSTALL_TYPE" == "CUSTOM" ]; then
-    printf '\n### Minimal install selected. No more dependencies to install.'
-elif [ "$INSTALL_TYPE" == "CUSTOM" ]; then
-    printf '\n### Installing custom-selected dependencies'
-    for opt in "${!opts[@]}"
-    do
-        if [[ ${opts[opt]} ]] ; then
-            if [ "$opt" == "1" ]; then
-                ${INSTALL_CMD} install-pip-dependency Adafruit_ADS1x15
-            elif [ "$opt" == "2" ]; then
-                ${INSTALL_DIRECTORY}/env/bin/pip3 install -e git://github.com/adafruit/Adafruit_Python_BME280.git#egg=adafruit-bme280
-            elif [ "$opt" == "3" ]; then
-                ${INSTALL_CMD} install-pip-dependency Adafruit_GPIO
-            elif [ "$opt" == "4" ]; then
-                ${INSTALL_CMD} install-pip-dependency Adafruit_MCP3008
-            elif [ "$opt" == "5" ]; then
-                ${INSTALL_CMD} install-pip-dependency Adafruit_TMP
-            elif [ "$opt" == "6" ]; then
-                ${INSTALL_CMD} install-pip-dependency MCP342x==0.3.3
-            elif [ "$opt" == "7" ]; then
-                ${INSTALL_CMD} install-pigpiod
-                ${INSTALL_CMD} enable-pigpiod-low
-            elif [ "$opt" == "8" ]; then
-                ${INSTALL_CMD} install-pip-dependency sht_sensor==17.5.5
-            elif [ "$opt" == "9" ]; then
-                ${INSTALL_CMD} install-pip-dependency tsl2561
-            elif [ "$opt" == "10" ]; then
-                ${INSTALL_DIRECTORY}/env/bin/pip3 install -e git://github.com/maxlklaxl/python-tsl2591.git#egg=tsl2591
-            elif [ "$opt" == "11" ]; then
-                ${INSTALL_CMD} install-pip-dependency w1thermsensor==1.0.5
+            option="${option%\"}"
+            option="${option#\"}"
+            if [ "$option" == "1" ]; then
+                ${INSTALL_DEP} Adafruit_ADS1x15 >>${LOG_LOCATION}
+            elif [ "$option" == "2" ]; then
+                ${INSTALL_DEP} Adafruit_Python_BME280 >>${LOG_LOCATION}
+            elif [ "$option" == "3" ]; then
+                ${INSTALL_DEP} Adafruit_GPIO >>${LOG_LOCATION}
+            elif [ "$option" == "4" ]; then
+                ${INSTALL_DEP} Adafruit_MCP3008 >>${LOG_LOCATION}
+            elif [ "$option" == "5" ]; then
+                ${INSTALL_DEP} Adafruit_TMP >>${LOG_LOCATION}
+            elif [ "$option" == "6" ]; then
+                ${INSTALL_DEP} MCP342x >>${LOG_LOCATION}
+            elif [ "$option" == "7" ]; then
+                ${INSTALL_DEP} install-pigpiod >>${LOG_LOCATION}
+            elif [ "$option" == "8" ]; then
+                ${INSTALL_DEP} sht_sensor >>${LOG_LOCATION}
+            elif [ "$option" == "9" ]; then
+                ${INSTALL_DEP} tsl2561 >>${LOG_LOCATION}
+            elif [ "$option" == "10" ]; then
+                ${INSTALL_DEP} tsl2591 >>${LOG_LOCATION}
+            elif [ "$option" == "11" ]; then
+                ${INSTALL_DEP} w1thermsensor >>${LOG_LOCATION}
             fi
-        fi
-    done
-elif [ "$INSTALL_TYPE" == "FULL" ]; then
-    ${INSTALL_CMD} install-pip-dependency Adafruit_ADS1x15
-    ${INSTALL_DIRECTORY}/env/bin/pip3 install -e git://github.com/adafruit/Adafruit_Python_BME280.git#egg=adafruit-bme280
-    ${INSTALL_CMD} install-pip-dependency Adafruit_GPIO
-    ${INSTALL_CMD} install-pip-dependency Adafruit_MCP3008
-    ${INSTALL_CMD} install-pip-dependency Adafruit_TMP
-    ${INSTALL_CMD} install-pip-dependency MCP342x==0.3.3
-    ${INSTALL_CMD} install-pigpiod
-    ${INSTALL_CMD} enable-pigpiod-low
-    ${INSTALL_CMD} install-pip-dependency sht_sensor==17.5.5
-    ${INSTALL_CMD} install-pip-dependency tsl2561
-    ${INSTALL_DIRECTORY}/env/bin/pip3 install -e git://github.com/maxlklaxl/python-tsl2591.git#egg=tsl2591
-    ${INSTALL_CMD} install-pip-dependency w1thermsensor==1.0.5
-fi
+        done
+    elif [ "$INSTALL_TYPE" == "full" ]; then
+        ${INSTALL_DEP} Adafruit_ADS1x15 >>${LOG_LOCATION}
+        ${INSTALL_DEP} Adafruit_Python_BME280 >>${LOG_LOCATION}
+        ${INSTALL_DEP} Adafruit_GPIO >>${LOG_LOCATION}
+        ${INSTALL_DEP} Adafruit_MCP3008 >>${LOG_LOCATION}
+        ${INSTALL_DEP} Adafruit_TMP >>${LOG_LOCATION}
+        ${INSTALL_DEP} MCP342x >>${LOG_LOCATION}
+        ${INSTALL_DEP} install-pigpiod >>${LOG_LOCATION}
+        ${INSTALL_DEP} sht_sensor >>${LOG_LOCATION}
+        ${INSTALL_DEP} tsl2561 >>${LOG_LOCATION}
+        ${INSTALL_DEP} tsl2591 >>${LOG_LOCATION}
+        ${INSTALL_DEP} w1thermsensor >>${LOG_LOCATION}
+    fi
 
-${INSTALL_CMD} update-influxdb
+    echo -e "XXX\n5\nInstalling InfluxDB... \nXXX"
+    ${INSTALL_CMD} update-influxdb >>${LOG_LOCATION}
 
-${INSTALL_CMD} update-influxdb-db-user
+    echo -e "XXX\n4\nInstalling InfluxDB database and user... \nXXX"
+    ${INSTALL_CMD} update-influxdb-db-user >>${LOG_LOCATION}
 
-${INSTALL_CMD} update-logrotate
+    echo -e "XXX\n5\nInstalling logrotate... \nXXX"
+    ${INSTALL_CMD} update-logrotate >>${LOG_LOCATION}
 
-${INSTALL_CMD} ssl-certs-generate
+    echo -e "XXX\n4\nGenerating SSL certificate... \nXXX"
+    ${INSTALL_CMD} ssl-certs-generate >>${LOG_LOCATION}
 
-${INSTALL_CMD} update-mycodo-startup-script
+    echo -e "XXX\n5\nInstalling Mycodo startup script... \nXXX"
+    ${INSTALL_CMD} update-mycodo-startup-script >>${LOG_LOCATION}
 
-${INSTALL_CMD} compile-translations
+    echo -e "XXX\n4\nCompiling translations... \nXXX"
+    ${INSTALL_CMD} compile-translations >>${LOG_LOCATION}
 
-${INSTALL_CMD} update-cron
+    echo -e "XXX\n5\nUpdating cron... \nXXX"
+    ${INSTALL_CMD} update-cron >>${LOG_LOCATION}
 
-${INSTALL_CMD} initialize
+    echo -e "XXX\n4\nInitializing install... \nXXX"
+    ${INSTALL_CMD} initialize >>${LOG_LOCATION}
 
-${INSTALL_CMD} web-server-update
+    echo -e "XXX\n5\nUpdating web server... \nXXX"
+    ${INSTALL_CMD} web-server-update >>${LOG_LOCATION}
 
-${INSTALL_CMD} web-server-restart
+    echo -e "XXX\n4\nRestarting web server... \nXXX"
+    ${INSTALL_CMD} web-server-restart >>${LOG_LOCATION}
 
-${INSTALL_CMD} web-server-connect
+    echo -e "XXX\n5\nConnecting to web server... \nXXX"
+    ${INSTALL_CMD} web-server-connect >>${LOG_LOCATION}
 
-${INSTALL_CMD} update-permissions
+    echo -e "XXX\n4\nUpdating permissions... \nXXX"
+    ${INSTALL_CMD} update-permissions >>${LOG_LOCATION}
 
-${INSTALL_CMD} restart-daemon
+    echo -e "XXX\n5\nRestarting daemon... \nXXX"
+    ${INSTALL_CMD} restart-daemon >>${LOG_LOCATION}
+
+} | whiptail --gauge "Installing Mycodo. Please wait..." 6 50 0
 
 trap : 0
 
@@ -247,12 +228,16 @@ if [[ -z ${IP} ]]; then
   IP="your.IP.address.here"
 fi
 
-date
-echo >&2 "
+CURRENT_DATE=$(date)
+printf "Mycodo Installer finished  ${CURRENT_DATE}" 2>&1 | tee -a ${LOG_LOCATION}
+echo "
 ************************************
 ** Mycodo successfully installed! **
 ************************************
 
+The full install log is located at:
+${INSTALL_DIRECTORY}/install/setup.log
+
 Go to https://${IP}/, or whatever your Raspberry Pi's
 IP address is, to create an admin user and log in.
-"
+" 2>&1 | tee -a ${LOG_LOCATION}